requirement.txt
# ======================
# some packages we have
# ======================
mysql-connector-python
bcrypt => for pw 
streamlit